draft_app/requirements.txt
db-dtypes==1.1.1
Flask==2.3.2
Flask-WTF==1.1.1
fsspec==2023.6.0
gcsfs==2023.6.0
google-auth==2.20.0
google-cloud-aiplatform==1.26.0
google-cloud-bigquery==3.11.1
google-cloud-storage==2.9.0
pandas==2.0.2